// src/app/types/workout.ts
export type Workout = {
  id: number;
  name: string;
  image: string;
  muscleGroups: string[];
  equipment: string;
  difficulty: string;
  duration: number;
  caloriesBurned: number;
  sets: number;
  reps: string;
  rating: number;
  description: string;
  instructions: string[];
};

export type SortKey = "duration" | "calories" | "rating";
export type PlanTab = "plan" | "saved";